M15_roles

+++ DatosSKD/Modulo15/ConexionBaseDatos.cs
@@ -227,16 +227,16 @@ namespace DatosSKD.Modulo15
                {
                    Implemento implemento = new Implemento();
                    implemento.Dojo_Implemento = new Dojo();
-                   implemento.Id_Implemento =    Convert .ToInt16(row[RecursosBDModulo15.tabla_idImplemento]);
+                   implemento.Id_Implemento =    Convert .ToInt32(row[RecursosBDModulo15.tabla_idImplemento]);
                    implemento.Nombre_Implemento =row[RecursosBDModulo15.tabla_nombreImplemento].ToString();
-                   implemento.Cantida_implemento = Convert.ToInt16(row[RecursosBDModulo15.tabla_cantidadImplemento]);
+                   implemento.Cantida_implemento = Convert.ToInt32(row[RecursosBDModulo15.tabla_cantidadImplemento]);
                    implemento.Imagen_implemento = row[RecursosBDModulo15.tabla_imagenImplemento].ToString();
                    implemento.Tipo_Implemento   =row[RecursosBDModulo15.tabla_tipoImplemento].ToString();
                    implemento.Marca_Implemento  =row[RecursosBDModulo15.tabla_marcaImplemento].ToString();
                    implemento.Color_Implemento  =row[RecursosBDModulo15.tabla_colorImplemento].ToString();
                    implemento.Talla_Implemento = row[RecursosBDModulo15.tabla_tallaImplemento].ToString();
-                   implemento.Dojo_Implemento.Dojo_Id =Convert.ToInt16(row[RecursosBDModulo15.tabla_dojoImplemento]);
-                   implemento.Stock_Minimo_Implemento = Convert.ToInt16(row[RecursosBDModulo15.tabla_stockImplemento]);
+                   implemento.Dojo_Implemento.Dojo_Id =Convert.ToInt32(row[RecursosBDModulo15.tabla_dojoImplemento]);
+                   implemento.Stock_Minimo_Implemento = Convert.ToInt32(row[RecursosBDModulo15.tabla_stockImplemento]);
                    implemento.Estatus_Implemento = row[RecursosBDModulo15.tabla_estatusImplemento].ToString();
                    implemento.Precio_Implemento =Convert.ToDouble(row[RecursosBDModulo15.tabla_precioImplemento]);
 
@@ -264,7 +264,6 @@ namespace DatosSKD.Modulo15
 
        }
        #endregion
-
 
        #region listarInventarioDatos2
        public static List<Implemento> listarInventarioDatos2(Dojo dojo)
@@ -388,7 +387,6 @@ namespace DatosSKD.Modulo15
 
        }
        #endregion
-
 
        #region eliminarInventarioDatos
        public  static void eliminarInventarioDatos(int idInventario,Dojo dojo)
